Закончил задачу по 5му стриму

+++ app/src/main/java/com/shiftsummer2020_2/screens/main/data/CitiesNetworkDataSourceImpl.kt
@@ -6,6 +6,6 @@ import exapmle.com.common.CityDto
 class CitiesNetworkDataSourceImpl(val api: CitiesApi): CitiesNetworkDataSource {
     override suspend fun getCities() = apiCallWrapper { api.getAll() }
     override suspend fun addCity(data: CityDto) = apiCallWrapper { api.addCity(data) }
-    override suspend fun deleteById(id: Int) = apiCallWrapper { api.deleteById(id) }
-    override suspend fun update(city: CityDto, id: Int) = apiCallWrapper { api.update(city, id) }
+    override suspend fun deleteById(id: Long) = apiCallWrapper { api.deleteById(id) }
+    override suspend fun update(city: CityDto, id: Long) = apiCallWrapper { api.update(city, id) }
 }
