+++ app/src/main/java/com/shiftsummer2020_2/screens/main/data/CitiesRepositoryImpl.kt
@@ -1,15 +1,13 @@
 package com.shiftsummer2020_2.screens.main.data
 
 import com.shiftsummer2020_2.screens.main.domain.CitiesRepository
-import com.shiftsummer2020_2.screens.main.domain.apiwrapper.ApiResultWrapper
-import exapmle.com.common.City
 import exapmle.com.common.CityDto
 
 class CitiesRepositoryImpl(private val citiesNetworkDataSource: CitiesNetworkDataSource): CitiesRepository {
     override suspend fun getCities() = citiesNetworkDataSource.getCities()
     override suspend fun addCity(city: CityDto) = citiesNetworkDataSource.addCity(city)
-    override suspend fun deleteCity(id: Int) = citiesNetworkDataSource.deleteById(id)
-    override suspend fun updateCity(city: CityDto, id: Int) = citiesNetworkDataSource.update(city, id)
+    override suspend fun deleteCity(id: Long) = citiesNetworkDataSource.deleteById(id)
+    override suspend fun updateCity(city: CityDto, id: Long) = citiesNetworkDataSource.update(city, id)
 
 
 }
