+++ app/src/main/java/com/shiftsummer2020_2/screens/main/domain/DeleteCityUseCase.kt
@@ -1,7 +1,5 @@
 package com.shiftsummer2020_2.screens.main.domain
 
-import exapmle.com.common.City
-
 class DeleteCityUseCase(private val citiesRepository: CitiesRepository) {
-    suspend operator fun invoke(id: Int) = citiesRepository.deleteCity(id)
+    suspend operator fun invoke(id: Long) = citiesRepository.deleteCity(id)
 }
